--- src/main/scala/dht/GetPeerKey.scala
@@ -1,13 +1,0 @@
-package dht
-
-object GetPeerKey {
-  /**
-   * user and location as a key of peers
-   *
-   * @param user     mail (not hashed)
-   * @param location location string
-   */
-  def apply(user: String, location: String): String = {
-    s"'${user}'@${location}"
-  }
-}
