+++ src/main/scala/logic/login/LocatorInfo.scala
@@ -1,6 +1,6 @@
-package login
+package logic.login
 
-import login.State.State
+import logic.login.State.State
 
 case class LocatorInfo(
                         locator: String, // a string defined by user, to tell on which machine the user is currently active
